result-set mappings
- re-enabled sq-query and resultset quoted-identifier tests
