--- hibernate-core/src/test/java/org/hibernate/orm/test/sql/hand/quotedidentifiers/Mappings.hbm.xml
@@ -1,41 +1,0 @@
-<?xml version="1.0"?>
-<!--
-  ~ Hibernate, Relational Persistence for Idiomatic Java
-  ~
-  ~ License: GNU Lesser General Public License (LGPL), version 2.1 or later.
-  ~ See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
-  -->
-<!DOCTYPE hibernate-mapping PUBLIC
-	"-//Hibernate/Hibernate Mapping DTD 3.0//EN"
-	"http://www.hibernate.org/dtd/hibernate-mapping-3.0.dtd">
-
-<hibernate-mapping package="org.hibernate.orm.test.sql.hand.quotedidentifiers">
-
-	<class name="Person" table="`Person`">
-		<id name="id" unsaved-value="0" column="`pId`">
-			<generator class="increment"/>
-		</id>
-		<property name="name" column="`pName`" not-null="true"/>
-	</class>
-    
-    <resultset name="person-scalar">
-        <return-scalar column="`pId`"/>
-        <return-scalar column="`pName`"/>
-    </resultset>
-
-    <resultset name="person-entity-basic">
-        <return alias="p" class="Person"/>
-    </resultset>
-
-    <resultset name="person-entity-expanded">
-		<return alias="p" class="Person">
-			<return-property name="id" column="`pId`"/>
-			<return-property name="name" column="`pName`"/>
-		</return>
-    </resultset>
-
-    <sql-query name="query-person">
-        select p."pId", p."pName" from "Person" p
-    </sql-query>
-	
-</hibernate-mapping>
