--- hibernate-core/src/test/java/org/hibernate/orm/test/sql/hand/quotedidentifiers/Person.java
@@ -1,41 +1,0 @@
-/*
- * Hibernate, Relational Persistence for Idiomatic Java
- *
- * License: GNU Lesser General Public License (LGPL), version 2.1 or later.
- * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
- */
-package org.hibernate.orm.test.sql.hand.quotedidentifiers;
-
-
-/**
- * TODO : javadoc
- *
- * @author Steve Ebersole
- */
-public class Person {
-	private Long id;
-	private String name;
-
-	public Person() {
-	}
-
-	public Person(String name) {
-		this.name = name;
-	}
-
-	public Long getId() {
-		return id;
-	}
-
-	public void setId(Long id) {
-		this.id = id;
-	}
-
-	public String getName() {
-		return name;
-	}
-
-	public void setName(String name) {
-		this.name = name;
-	}
-}
