+++ hibernate-core/src/test/resources/mappings/query/named/sql/hbm/QuotedIdentifiers.hbm.xml
@@ -1,0 +1,34 @@
+<?xml version="1.0"?>
+<!--
+  ~ Hibernate, Relational Persistence for Idiomatic Java
+  ~
+  ~ License: GNU Lesser General Public License (LGPL), version 2.1 or later
+  ~ See the lgpl.txt file in the root directory or http://www.gnu.org/licenses/lgpl-2.1.html
+  -->
+<!DOCTYPE hibernate-mapping PUBLIC
+	"-//Hibernate/Hibernate Mapping DTD 3.0//EN"
+	"http://www.hibernate.org/dtd/hibernate-mapping-3.0.dtd">
+
+<hibernate-mapping package="org.hibernate.orm.test.query.sql.named" default-access="field">
+
+    <resultset name="person-scalar">
+        <return-scalar column="`id`"/>
+        <return-scalar column="`name`"/>
+    </resultset>
+
+    <resultset name="person-entity-basic">
+        <return alias="p" class="QuotedPerson"/>
+    </resultset>
+
+    <resultset name="person-entity-expanded">
+		<return alias="p" class="QuotedPerson">
+			<return-property name="id" column="`id`"/>
+			<return-property name="name" column="`name`"/>
+		</return>
+    </resultset>
+
+    <sql-query name="query-person">
+        select p."id", p."name" from "Person" p
+    </sql-query>
+	
+</hibernate-mapping>
